Cut wire-byte load + add cache tiers (post-stress-test fixes)
The 27.5-min k6 stress test (2,000 VUs, 340K requests) showed
http_req_receiving p95 = 4.4s — response transfer was the dominant
latency contributor, not server thinking time (waiting p95 = 771ms).
Two attack surfaces: oversized JSON payloads, and no CDN caching.

lib/api.ts — four-tier cache vocabulary
- "live"   max-age=10,  s-maxage=60,   stale-while-revalidate=120
- "short"  max-age=30,  s-maxage=60,   stale-while-revalidate=300  (default)
- "long"   max-age=120, s-maxage=600,  stale-while-revalidate=1800
- "static" max-age=300, s-maxage=3600, stale-while-revalidate=86400
Each tier now carries BOTH max-age (browser) and s-maxage (CDN) so
caching works whether or not we put a CDN in front. Added Vary:
Accept-Encoding so any intermediate cache keeps gzip+brotli+identity
variants separate.

Cache tier assignments
- "static" : /api/quarters, /api/best-quarter, /api/commentary,
             /api/expert-quotes — basically never change between requests
- "long"   : /api/related-coverage — refreshed 3× daily by the cron job
- "live"   : /api/market-context, /api/prices — refresh every minute
- "short"  : everything else (sectors, summary, dashboard, upcoming,
             q4-announcements, trends, quarterly-results, companies/[ticker])
- "no-store": /api/refresh-company (write-like side-effect endpoint)

Payload trimming (drops select("*") to projected columns)
- /api/companies: 6 cols instead of every metadata field (~60% smaller
  payload on the 200-row response)
- /api/companies/[ticker]: 10 cols, drops raw_json + updated_at
- /api/dashboard: companies select narrowed to 8 cols, drops raw_json
  + ingestion-only fields that never reach the UI
- /api/quarterly-results (single-ticker branch): explicit column list,
  drops raw_json (the bulkiest field on quarterly_financials rows)

next.config.js
- compress: true set explicitly (default in Next 13+, but pinning it so
  it can't be silently turned off). With Next's standalone server doing
  gzip on the way out, JSON payloads shrink ~75% — directly addressing
  the http_req_receiving p95 in the stress test.

Build verified, all 23 routes compile clean.

+++ app/api/related-coverage/route.ts
@@ -26,5 +26,8 @@ export async function GET(_req: NextRequest) {
     .limit(20);
 
   if (error) return jsonError(error.message, 500);
-  return jsonOk(data ?? [], { cache: "short" });
+  // Coverage rows are populated 3× daily by fetch_related_coverage.py —
+  // long cache tier (10 min edge / 2 min browser) matches that cadence
+  // without ever serving truly stale data.
+  return jsonOk(data ?? [], { cache: "long" });
 }
